feat: move projection head to tailor
