+++ finetuner/tuner/keras/__init__.py
@@ -205,43 +205,3 @@ def get_device(device: str):
     elif device == 'cpu':
         device = '/CPU:0'
     return tf.device(device)
-
-
-class _ProjectionHead(tf.keras.layers.Layer):
-    """Projection head used internally for self-supervised training.
-    It is (by default) a simple 3-layer MLP to be attached on top of embedding model only for training purpose.
-    After training, it should be cut-out from the embedding model.
-    """
-
-    EPSILON = 1e-5
-
-    def __init__(self, in_features: int, output_dim: int = 128, num_layers: int = 3):
-        super().__init__()
-        self.head_layers = []
-        is_last_layer = False
-        for idx in range(num_layers):
-            if idx == num_layers - 1:
-                is_last_layer = True
-            if not is_last_layer:
-                self.head_layers.append(
-                    tf.keras.layers.Dense(
-                        units=in_features,
-                        bias_initializer='zeros',
-                    )
-                )
-                self.head_layers.append(
-                    tf.keras.layers.BatchNormalization(epsilon=self.EPSILON)
-                )
-                self.head_layers.append(tf.keras.layers.ReLU())
-            else:
-                self.head_layers.append(
-                    tf.keras.layers.Dense(
-                        units=output_dim,
-                        bias_initializer='zeros',
-                    )
-                )
-
-    def call(self, x):
-        for layer in self.head_layers:
-            x = layer(x)
-        return x
